refactor(commands): split projections.rs (824 lines) into submodules
U-16 (<=800 lines/file) was violated (by 24 lines). Move registry
projection helpers, skill inventory, remote status payload, and
rollback utilities into focused submodules.

- projections/mod.rs    129
- registry.rs           397
- inventory.rs          102
- remote.rs             197
- rollback.rs            46
