--- src/commands/projections.rs
@@ -1,824 +1,0 @@
-use std::collections::BTreeSet;
-use std::fs;
-use std::path::{Path, PathBuf};
-
-use anyhow::{Context, Result};
-use chrono::Utc;
-use serde_json::json;
-use uuid::Uuid;
-
-use crate::cli::{CaptureArgs, ProjectionMethod};
-use crate::envelope::Meta;
-use crate::gitops;
-use crate::state::{AppContext, PendingOpsReport, resolve_agent_skill_source_dirs};
-use crate::state_model::{
-    RegistryBindingRule, RegistryObservationEvent, RegistryOperationRecord,
-    RegistryProjectionInstance, RegistryProjectionsFile, RegistryRulesFile, RegistrySnapshot,
-    RegistryStatePaths,
-};
-use crate::types::{ErrorCode, SyncState};
-
-use super::CommandFailure;
-use super::event_store::redact_sensitive_string;
-use super::file_ops::{
-    copy_dir_recursive, copy_dir_recursive_preserving_symlinks, create_symlink_dir,
-};
-use super::helpers::{map_git, map_io, map_push_rejected, map_queue, map_remote_unreachable};
-use crate::state::remove_path_if_exists;
-
-// ---------------------------------------------------------------------------
-// Registry state mutators
-// ---------------------------------------------------------------------------
-
-pub(crate) fn upsert_rule(rules: &mut RegistryRulesFile, rule: RegistryBindingRule) {
-    if let Some(existing) = rules.rules.iter_mut().find(|existing| {
-        existing.binding_id == rule.binding_id
-            && existing.skill_id == rule.skill_id
-            && existing.target_id == rule.target_id
-    }) {
-        existing.method = rule.method;
-        existing.watch_policy = rule.watch_policy;
-        return;
-    }
-
-    rules.rules.push(rule);
-    rules.rules.sort_by(|left, right| {
-        left.binding_id
-            .cmp(&right.binding_id)
-            .then_with(|| left.skill_id.cmp(&right.skill_id))
-            .then_with(|| left.target_id.cmp(&right.target_id))
-    });
-}
-
-pub(crate) fn upsert_projection(
-    projections: &mut RegistryProjectionsFile,
-    projection: RegistryProjectionInstance,
-) {
-    if let Some(existing) = projections
-        .projections
-        .iter_mut()
-        .find(|existing| existing.instance_id == projection.instance_id)
-    {
-        *existing = projection;
-        return;
-    }
-
-    projections.projections.push(projection);
-    projections
-        .projections
-        .sort_by(|left, right| left.instance_id.cmp(&right.instance_id));
-}
-
-pub(crate) fn project_skill_to_target(
-    src: &Path,
-    dst: &Path,
-    method: ProjectionMethod,
-) -> Result<()> {
-    match method {
-        ProjectionMethod::Symlink => create_symlink_dir(src, dst),
-        ProjectionMethod::Copy => {
-            let parent = dst
-                .parent()
-                .context("projection target has no parent directory")?;
-            let tmp_dir = parent.join(format!(".loom-tmp-{}", Uuid::new_v4()));
-            if let Err(err) = copy_dir_recursive_preserving_symlinks(src, &tmp_dir) {
-                let _ = remove_path_if_exists(&tmp_dir);
-                return Err(err);
-            }
-            if let Err(err) = std::fs::rename(&tmp_dir, dst) {
-                let _ = remove_path_if_exists(&tmp_dir);
-                return Err(err).context("failed to atomically place projection");
-            }
-            Ok(())
-        }
-        ProjectionMethod::Materialize => {
-            let parent = dst
-                .parent()
-                .context("projection target has no parent directory")?;
-            let tmp_dir = parent.join(format!(".loom-tmp-{}", Uuid::new_v4()));
-            if let Err(err) = copy_dir_recursive(src, &tmp_dir) {
-                let _ = remove_path_if_exists(&tmp_dir);
-                return Err(err);
-            }
-            if let Err(err) = std::fs::rename(&tmp_dir, dst) {
-                let _ = remove_path_if_exists(&tmp_dir);
-                return Err(err).context("failed to atomically place projection");
-            }
-            Ok(())
-        }
-    }
-}
-
-pub(crate) fn resolve_capture_projection(
-    snapshot: &RegistrySnapshot,
-    args: &CaptureArgs,
-) -> std::result::Result<RegistryProjectionInstance, CommandFailure> {
-    if let Some(instance_id) = args.instance.as_deref() {
-        let projection = snapshot
-            .projections
-            .projections
-            .iter()
-            .find(|projection| projection.instance_id == instance_id)
-            .cloned()
-            .ok_or_else(|| {
-                CommandFailure::new(
-                    ErrorCode::ArgInvalid,
-                    format!("projection instance '{}' not found", instance_id),
-                )
-            })?;
-        if let Some(skill) = args.skill.as_deref()
-            && projection.skill_id != skill
-        {
-            return Err(CommandFailure::new(
-                ErrorCode::ArgInvalid,
-                format!(
-                    "instance '{}' belongs to skill '{}' not '{}'",
-                    instance_id, projection.skill_id, skill
-                ),
-            ));
-        }
-        if let Some(expected_binding) = args.binding.as_deref()
-            && projection.binding_id.as_deref() != Some(expected_binding)
-        {
-            return Err(CommandFailure::new(
-                ErrorCode::ArgInvalid,
-                format!(
-                    "instance '{}' belongs to binding '{}' not '{}'",
-                    instance_id,
-                    projection.binding_id.as_deref().unwrap_or("(orphaned)"),
-                    expected_binding
-                ),
-            ));
-        }
-        return Ok(projection);
-    }
-
-    let skill = args.skill.as_deref().ok_or_else(|| {
-        CommandFailure::new(
-            ErrorCode::ArgInvalid,
-            "capture requires <skill> or --instance",
-        )
-    })?;
-    let binding_id = args.binding.as_deref().ok_or_else(|| {
-        CommandFailure::new(
-            ErrorCode::ArgInvalid,
-            "capture requires --binding when --instance is not provided",
-        )
-    })?;
-
-    let matches = snapshot
-        .projections
-        .projections
-        .iter()
-        .filter(|projection| {
-            projection.skill_id == skill && projection.binding_id.as_deref() == Some(binding_id)
-        })
-        .cloned()
-        .collect::<Vec<_>>();
-
-    match matches.len() {
-        0 => Err(CommandFailure::new(
-            ErrorCode::ArgInvalid,
-            format!(
-                "no projection found for skill '{}' and binding '{}'",
-                skill, binding_id
-            ),
-        )),
-        1 => Ok(matches.into_iter().next().expect("single projection")),
-        _ => Err(CommandFailure::new(
-            ErrorCode::ArgInvalid,
-            format!(
-                "multiple projections found for skill '{}' and binding '{}'; use --instance",
-                skill, binding_id
-            ),
-        )),
-    }
-}
-
-pub(crate) fn update_projection_after_capture(
-    projections: &mut RegistryProjectionsFile,
-    instance_id: &str,
-    rev: &str,
-) -> std::result::Result<(), CommandFailure> {
-    let projection = projections
-        .projections
-        .iter_mut()
-        .find(|projection| projection.instance_id == instance_id)
-        .ok_or_else(|| {
-            CommandFailure::new(
-                ErrorCode::ArgInvalid,
-                format!(
-                    "projection instance '{}' not found during capture update",
-                    instance_id
-                ),
-            )
-        })?;
-    projection.last_applied_rev = rev.to_string();
-    projection.health = "healthy".to_string();
-    projection.observed_drift = Some(false);
-    projection.updated_at = Some(Utc::now());
-    Ok(())
-}
-
-pub(crate) fn record_registry_operation(
-    paths: &RegistryStatePaths,
-    intent: &str,
-    payload: serde_json::Value,
-    effects: serde_json::Value,
-) -> Result<String> {
-    let op_id = format!("op_{}", Uuid::new_v4().simple());
-    let now = Utc::now();
-    let record = RegistryOperationRecord {
-        op_id: op_id.clone(),
-        intent: intent.to_string(),
-        status: "succeeded".to_string(),
-        ack: false,
-        payload,
-        effects,
-        last_error: None,
-        created_at: now,
-        updated_at: now,
-    };
-    let operations_len = fs::metadata(&paths.operations_file)
-        .with_context(|| {
-            format!(
-                "failed to stat operations log {} before append",
-                paths.operations_file.display()
-            )
-        })?
-        .len();
-    let checkpoint_backup = fs::read(&paths.checkpoint_file).with_context(|| {
-        format!(
-            "failed to snapshot checkpoint {} before operation append",
-            paths.checkpoint_file.display()
-        )
-    })?;
-
-    let persist_result: Result<()> = (|| -> Result<()> {
-        paths.append_operation(&record)?;
-        maybe_projection_fault("record_v3_operation_after_append")?;
-
-        let mut checkpoint = paths.load_checkpoint()?;
-        checkpoint.last_scanned_op_id = Some(op_id.clone());
-        checkpoint.updated_at = now;
-        paths.save_checkpoint(&checkpoint)?;
-        maybe_projection_fault("record_v3_operation_after_checkpoint")?;
-        Ok(())
-    })();
-
-    if let Err(err) = persist_result {
-        if let Err(rollback_err) =
-            rollback_record_registry_operation(paths, operations_len, &checkpoint_backup)
-        {
-            return Err(err.context(format!(
-                "failed to rollback registry operation record after partial write: {}",
-                rollback_err
-            )));
-        }
-        return Err(err);
-    }
-
-    Ok(op_id)
-}
-
-pub(crate) fn record_registry_observation(
-    paths: &RegistryStatePaths,
-    instance_id: &str,
-    kind: &str,
-    path: Option<String>,
-    from: Option<String>,
-    to: Option<String>,
-) -> Result<String> {
-    let event_id = Uuid::new_v4().to_string();
-    let event = RegistryObservationEvent {
-        event_id: event_id.clone(),
-        instance_id: instance_id.to_string(),
-        kind: kind.to_string(),
-        path,
-        from,
-        to,
-        observed_at: Utc::now(),
-    };
-    paths.append_observation(&event)?;
-    Ok(event_id)
-}
-
-#[derive(Debug, Clone)]
-pub(crate) struct RegistryAuditStateBackup {
-    operations: Vec<u8>,
-    checkpoint: Vec<u8>,
-    observations: Vec<(String, Vec<u8>)>,
-}
-
-pub(crate) fn snapshot_registry_audit_state(
-    paths: &RegistryStatePaths,
-) -> Result<RegistryAuditStateBackup> {
-    let mut observations = Vec::new();
-    if paths.observations_dir.exists() {
-        for entry in fs::read_dir(&paths.observations_dir).with_context(|| {
-            format!(
-                "failed to read observations dir {}",
-                paths.observations_dir.display()
-            )
-        })? {
-            let entry = entry.with_context(|| {
-                format!(
-                    "failed to read observations entry under {}",
-                    paths.observations_dir.display()
-                )
-            })?;
-            let file_type = entry.file_type().with_context(|| {
-                format!(
-                    "failed to inspect observation entry {}",
-                    entry.path().display()
-                )
-            })?;
-            if !file_type.is_file() {
-                continue;
-            }
-            let name = entry.file_name().to_string_lossy().into_owned();
-            let contents = fs::read(entry.path()).with_context(|| {
-                format!("failed to snapshot observation {}", entry.path().display())
-            })?;
-            observations.push((name, contents));
-        }
-    }
-
-    Ok(RegistryAuditStateBackup {
-        operations: fs::read(&paths.operations_file)
-            .with_context(|| format!("failed to snapshot {}", paths.operations_file.display()))?,
-        checkpoint: fs::read(&paths.checkpoint_file)
-            .with_context(|| format!("failed to snapshot {}", paths.checkpoint_file.display()))?,
-        observations,
-    })
-}
-
-pub(crate) fn restore_registry_audit_state(
-    paths: &RegistryStatePaths,
-    backup: &RegistryAuditStateBackup,
-) -> Result<()> {
-    fs::write(&paths.operations_file, &backup.operations)
-        .with_context(|| format!("failed to restore {}", paths.operations_file.display()))?;
-    fs::write(&paths.checkpoint_file, &backup.checkpoint)
-        .with_context(|| format!("failed to restore {}", paths.checkpoint_file.display()))?;
-
-    fs::create_dir_all(&paths.observations_dir).with_context(|| {
-        format!(
-            "failed to create observations dir {}",
-            paths.observations_dir.display()
-        )
-    })?;
-    for entry in fs::read_dir(&paths.observations_dir).with_context(|| {
-        format!(
-            "failed to read observations dir {}",
-            paths.observations_dir.display()
-        )
-    })? {
-        let entry = entry.with_context(|| {
-            format!(
-                "failed to read observations entry under {}",
-                paths.observations_dir.display()
-            )
-        })?;
-        let path = entry.path();
-        let file_type = entry
-            .file_type()
-            .with_context(|| format!("failed to inspect observation entry {}", path.display()))?;
-        if file_type.is_dir() {
-            fs::remove_dir_all(&path)
-                .with_context(|| format!("failed to remove observation dir {}", path.display()))?;
-        } else {
-            fs::remove_file(&path)
-                .with_context(|| format!("failed to remove observation file {}", path.display()))?;
-        }
-    }
-    for (name, contents) in &backup.observations {
-        let path = paths.observations_dir.join(name);
-        fs::write(&path, contents)
-            .with_context(|| format!("failed to restore observation {}", path.display()))?;
-    }
-    Ok(())
-}
-
-fn maybe_projection_fault(tag: &str) -> Result<()> {
-    if std::env::var("LOOM_FAULT_INJECT").ok().as_deref() == Some(tag) {
-        return Err(anyhow::anyhow!("fault injected at {}", tag));
-    }
-    Ok(())
-}
-
-fn rollback_record_registry_operation(
-    paths: &RegistryStatePaths,
-    operations_len: u64,
-    checkpoint_backup: &[u8],
-) -> Result<()> {
-    truncate_file(&paths.operations_file, operations_len)?;
-    restore_raw_file(&paths.checkpoint_file, checkpoint_backup)?;
-    Ok(())
-}
-
-fn truncate_file(path: &Path, len: u64) -> Result<()> {
-    let file = fs::OpenOptions::new()
-        .write(true)
-        .open(path)
-        .with_context(|| format!("failed to open file for rollback {}", path.display()))?;
-    file.set_len(len)
-        .with_context(|| format!("failed to truncate file {}", path.display()))?;
-    file.sync_all()
-        .with_context(|| format!("failed to sync truncated file {}", path.display()))?;
-    Ok(())
-}
-
-fn restore_raw_file(path: &Path, contents: &[u8]) -> Result<()> {
-    let parent = path
-        .parent()
-        .context("cannot restore file without parent directory")?;
-    fs::create_dir_all(parent)
-        .with_context(|| format!("failed to create restore dir {}", parent.display()))?;
-    fs::write(path, contents)
-        .with_context(|| format!("failed to restore file {}", path.display()))?;
-    Ok(())
-}
-
-// ---------------------------------------------------------------------------
-// SkillInventory
-// ---------------------------------------------------------------------------
-
-#[derive(Debug, Clone, Default)]
-pub struct SkillInventory {
-    pub source_skills: Vec<String>,
-    pub backup_skills: Vec<String>,
-    pub source_dirs: Vec<PathBuf>,
-    pub warnings: Vec<String>,
-}
-
-pub fn collect_skill_inventory(ctx: &AppContext) -> SkillInventory {
-    let source_dirs = resolve_agent_skill_source_dirs(&ctx.root);
-    let mut warnings = Vec::new();
-
-    let source_skills = list_unique_skills_from_dirs(&source_dirs, "source", &mut warnings);
-    let backup_skills = list_unique_skills_from_dirs(
-        std::slice::from_ref(&ctx.skills_dir),
-        "backup",
-        &mut warnings,
-    );
-
-    SkillInventory {
-        source_skills,
-        backup_skills,
-        source_dirs,
-        warnings,
-    }
-}
-
-fn list_unique_skills_from_dirs(
-    dirs: &[PathBuf],
-    label: &str,
-    warnings: &mut Vec<String>,
-) -> Vec<String> {
-    let mut skills = BTreeSet::new();
-
-    for dir in dirs {
-        if !dir.exists() {
-            continue;
-        }
-
-        let entries = match fs::read_dir(dir) {
-            Ok(entries) => entries,
-            Err(err) => {
-                warnings.push(format!(
-                    "failed to read {} skills dir {}: {}",
-                    label,
-                    dir.display(),
-                    err
-                ));
-                continue;
-            }
-        };
-
-        for entry in entries {
-            let entry = match entry {
-                Ok(entry) => entry,
-                Err(err) => {
-                    warnings.push(format!(
-                        "failed to read entry in {} skills dir {}: {}",
-                        label,
-                        dir.display(),
-                        err
-                    ));
-                    continue;
-                }
-            };
-
-            let is_dir = match entry.file_type() {
-                Ok(kind) if kind.is_dir() => true,
-                Ok(kind) if kind.is_symlink() => fs::metadata(entry.path())
-                    .map(|meta| meta.is_dir())
-                    .unwrap_or(false),
-                Ok(_) => false,
-                Err(err) => {
-                    warnings.push(format!(
-                        "failed to inspect entry {} in {} skills dir {}: {}",
-                        entry.file_name().to_string_lossy(),
-                        label,
-                        dir.display(),
-                        err
-                    ));
-                    false
-                }
-            };
-
-            if is_dir {
-                skills.insert(entry.file_name().to_string_lossy().to_string());
-            }
-        }
-    }
-
-    skills.into_iter().collect()
-}
-
-// ---------------------------------------------------------------------------
-// Remote status / sync internals
-// ---------------------------------------------------------------------------
-
-pub fn remote_status_payload(
-    ctx: &AppContext,
-) -> std::result::Result<(serde_json::Value, Meta), CommandFailure> {
-    let pending_report = ctx.read_pending_report().map_err(map_io)?;
-    remote_status_payload_with_pending(ctx, pending_report)
-}
-
-pub(crate) fn remote_status_payload_with_pending(
-    ctx: &AppContext,
-    pending_report: PendingOpsReport,
-) -> std::result::Result<(serde_json::Value, Meta), CommandFailure> {
-    let pending = pending_report.ops.len();
-
-    if !gitops::remote_exists(ctx) {
-        return Ok((
-            json!({
-                "configured": false,
-                "pending_ops": pending,
-                "sync_state": SyncState::LocalOnly,
-            }),
-            Meta {
-                warnings: pending_report
-                    .warnings
-                    .into_iter()
-                    .chain(std::iter::once("remote origin not configured".to_string()))
-                    .collect(),
-                sync_state: Some(SyncState::LocalOnly),
-                op_id: None,
-            },
-        ));
-    }
-
-    let url = gitops::remote_url(ctx)
-        .map_err(map_git)?
-        .unwrap_or_default();
-    let redacted_url = redact_sensitive_string(&url);
-    let mut meta = Meta {
-        warnings: pending_report.warnings,
-        sync_state: None,
-        op_id: None,
-    };
-
-    if !gitops::remote_tracking_main_exists(ctx).map_err(map_git)? {
-        let sync_state = if pending > 0 {
-            SyncState::PendingPush
-        } else {
-            SyncState::LocalOnly
-        };
-        meta.warnings.push(
-            "origin/main has not been fetched yet; status is based on local state".to_string(),
-        );
-        meta.sync_state = Some(sync_state.clone());
-        return Ok((
-            json!({
-                "configured": true,
-                "remote": "origin",
-                "url": redacted_url,
-                "pending_ops": pending,
-                "tracking_ref": false,
-                "sync_state": sync_state,
-            }),
-            meta,
-        ));
-    }
-
-    let (ahead, behind) = gitops::ahead_behind_main(ctx).map_err(map_git)?;
-    let sync_state = if pending > 0 {
-        SyncState::PendingPush
-    } else if ahead == 0 && behind == 0 {
-        SyncState::Synced
-    } else if ahead > 0 && behind == 0 {
-        SyncState::PendingPush
-    } else {
-        SyncState::Diverged
-    };
-    meta.sync_state = Some(sync_state.clone());
-
-    Ok((
-        json!({
-            "configured": true,
-            "remote": "origin",
-            "url": redacted_url,
-            "ahead": ahead,
-            "behind": behind,
-            "pending_ops": pending,
-            "tracking_ref": true,
-            "sync_state": sync_state,
-        }),
-        meta,
-    ))
-}
-
-pub(crate) fn maybe_autosync_or_queue(
-    ctx: &AppContext,
-    command: &str,
-    request_id: &str,
-    details: serde_json::Value,
-    meta: &mut Meta,
-) -> std::result::Result<(), CommandFailure> {
-    if !gitops::remote_exists(ctx) {
-        ctx.append_pending(command, details, request_id.to_string())
-            .map_err(map_queue)?;
-        meta.sync_state = Some(SyncState::PendingPush);
-        meta.warnings
-            .push("remote origin not configured, operation queued".to_string());
-        return Ok(());
-    }
-
-    match sync_push_internal(ctx) {
-        Ok(_) => {
-            meta.sync_state = Some(SyncState::Synced);
-        }
-        Err(err) => {
-            ctx.append_pending(command, details, request_id.to_string())
-                .map_err(map_queue)?;
-            meta.sync_state = Some(match err.code {
-                ErrorCode::RemoteDiverged => SyncState::Diverged,
-                ErrorCode::ReplayConflict => SyncState::Conflicted,
-                _ => SyncState::PendingPush,
-            });
-            meta.warnings.push(format!(
-                "auto sync failed ({}), operation queued",
-                err.code.as_str()
-            ));
-        }
-    }
-    Ok(())
-}
-
-pub(crate) fn sync_push_internal(
-    ctx: &AppContext,
-) -> std::result::Result<&'static str, CommandFailure> {
-    if !gitops::remote_exists(ctx) {
-        return Err(CommandFailure::new(
-            ErrorCode::ArgInvalid,
-            "remote origin not configured",
-        ));
-    }
-
-    let _state_commit = gitops::commit_paths_if_changed(
-        ctx,
-        &[".gitignore", "state/registry", "state/v3"],
-        "sync: commit registry state",
-    )
-    .map_err(map_git)?;
-    let pending_report = ctx.read_pending_report().map_err(map_io)?;
-    let queued_ids = pending_report
-        .ops
-        .iter()
-        .map(|op| op.stable_id())
-        .collect::<std::collections::BTreeSet<_>>();
-    let remote_main_exists =
-        gitops::fetch_origin_main_if_present(ctx).map_err(map_remote_unreachable)?;
-    let remote_history_exists =
-        gitops::fetch_origin_history_branch_if_present(ctx).map_err(map_remote_unreachable)?;
-    if remote_history_exists {
-        let _ = gitops::sync_history_branch_from_remote(ctx).map_err(map_git)?;
-    }
-    if remote_main_exists {
-        let (_ahead, behind) = gitops::ahead_behind_main(ctx).map_err(map_git)?;
-        if behind > 0 {
-            return Err(CommandFailure::new(
-                ErrorCode::RemoteDiverged,
-                "local branch is behind origin/main",
-            ));
-        }
-    }
-    gitops::push_main_with_tags(ctx).map_err(map_push_rejected)?;
-    ctx.remove_pending_ops(&queued_ids).map_err(map_queue)?;
-    Ok("pushed")
-}
-
-pub(crate) fn sync_replay_internal(
-    ctx: &AppContext,
-) -> std::result::Result<&'static str, CommandFailure> {
-    let pending = ctx.pending_count().map_err(map_io)?;
-    if pending == 0 {
-        return Ok("no_pending_ops");
-    }
-    sync_push_internal(ctx)?;
-    Ok("replayed")
-}
-
-#[cfg(test)]
-mod project_skill_tests {
-    use super::*;
-    use std::env;
-    use std::fs as stdfs;
-
-    fn scratch_dir(label: &str) -> PathBuf {
-        let dir = env::temp_dir().join(format!(
-            "loom-projections-{}-{}",
-            label,
-            Uuid::new_v4().simple()
-        ));
-        stdfs::create_dir_all(&dir).expect("scratch dir");
-        dir
-    }
-
-    fn make_skill_src(base: &Path) -> PathBuf {
-        let skill = base.join("sample-skill");
-        stdfs::create_dir_all(&skill).expect("skill dir");
-        stdfs::write(skill.join("SKILL.md"), "# sample\n").expect("SKILL.md");
-        skill
-    }
-
-    #[test]
-    fn copy_method_materializes_files() {
-        let base = scratch_dir("copy");
-        let src = make_skill_src(&base);
-        let dst = base.join("dst-copy");
-        project_skill_to_target(&src, &dst, ProjectionMethod::Copy).expect("copy ok");
-        assert!(dst.join("SKILL.md").is_file(), "SKILL.md must be copied");
-        let _ = stdfs::remove_dir_all(&base);
-    }
-
-    #[test]
-    fn materialize_method_materializes_files() {
-        let base = scratch_dir("materialize");
-        let src = make_skill_src(&base);
-        let dst = base.join("dst-mat");
-        project_skill_to_target(&src, &dst, ProjectionMethod::Materialize).expect("materialize ok");
-        assert!(dst.join("SKILL.md").is_file());
-        let _ = stdfs::remove_dir_all(&base);
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn copy_preserves_symlink_but_materialize_resolves_it() {
-        let base = scratch_dir("copy-vs-materialize-symlink");
-        let src = make_skill_src(&base);
-        let secret = base.join("secret.txt");
-        stdfs::write(&secret, "secret contents\n").expect("secret file");
-        std::os::unix::fs::symlink(&secret, src.join("secret-link")).expect("source symlink");
-
-        let copy_dst = base.join("dst-copy");
-        project_skill_to_target(&src, &copy_dst, ProjectionMethod::Copy).expect("copy ok");
-        assert!(
-            stdfs::symlink_metadata(copy_dst.join("secret-link"))
-                .expect("copy link metadata")
-                .file_type()
-                .is_symlink(),
-            "copy must preserve the symlink instead of dereferencing it"
-        );
-
-        let mat_dst = base.join("dst-mat");
-        project_skill_to_target(&src, &mat_dst, ProjectionMethod::Materialize)
-            .expect("materialize ok");
-        assert!(
-            stdfs::symlink_metadata(mat_dst.join("secret-link"))
-                .expect("materialized link metadata")
-                .is_file(),
-            "materialize must produce a real file"
-        );
-        assert_eq!(
-            stdfs::read_to_string(mat_dst.join("secret-link")).expect("materialized content"),
-            "secret contents\n"
-        );
-
-        let _ = stdfs::remove_dir_all(&base);
-    }
-
-    #[test]
-    fn symlink_method_creates_link_on_unix_tmp() {
-        if !cfg!(unix) {
-            return;
-        }
-        let base = scratch_dir("symlink");
-        let src = make_skill_src(&base);
-        let dst = base.join("dst-symlink");
-        project_skill_to_target(&src, &dst, ProjectionMethod::Symlink).expect("symlink ok");
-        assert!(
-            stdfs::symlink_metadata(&dst)
-                .expect("dst exists")
-                .file_type()
-                .is_symlink(),
-            "dst must be a symlink"
-        );
-        let _ = stdfs::remove_dir_all(&base);
-    }
-}
